Moved backend test to specific file
